LPS-150306 Call delete segments experiences only if we have, since we are sharing the experiences between draft and published page

+++ modules/apps/layout/layout-page-template-service/src/main/java/com/liferay/layout/page/template/internal/model/listener/LayoutModelListener.java
@@ -126,9 +126,16 @@ public class LayoutModelListener extends BaseModelListener<Layout> {
 		}
 
 		try {
-			_segmentsExperienceLocalService.deleteSegmentsExperiences(
-				layout.getGroupId(), _portal.getClassNameId(Layout.class),
-				layout.getPlid());
+			int count =
+				_segmentsExperienceLocalService.getSegmentsExperiencesCount(
+					layout.getGroupId(), _portal.getClassNameId(Layout.class),
+					layout.getPlid());
+
+			if (count > 0) {
+				_segmentsExperienceLocalService.deleteSegmentsExperiences(
+					layout.getGroupId(), _portal.getClassNameId(Layout.class),
+					layout.getPlid());
+			}
 
 			if (!layout.isTypeContent()) {
 				return;
